add the mochicow_http_protocol
cowboy keep a buffer around when it parses the request using the
`cowboy_http_protocol` which mqke it imposible to use with mochiweb
since there is no way to pass this buffer correctly to the mochiweb
request. `mochicow_http_protocol` just works like this
protocol except it doesn't use a buffer.