s/mochicow_http_protocol/mochicow_request
Big change in teh way a mochizeb request is now handled. Previous
attempt was removing the use of the buffer. But this change was
problematic for some cowboy handlers. Instead now we use our own
mochieb request module. The change is transparent for the users.

This module keep a receive buffer around and pass it to the cowby
protocol zhen it ends. Tested on rcouch and all JS tests passed.